# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Krutrim Agent Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chats": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create Chat",
        "operationId": "create_chat_api_chats_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "List Chats",
        "operationId": "list_chats_api_chats_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Chat"
                  },
                  "title": "Response List Chats Api Chats Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "Get Chat",
        "operationId": "get_chat_api_chats__chat_id__get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "chats"
        ],
        "summary": "Update Chat",
        "operationId": "update_chat_api_chats__chat_id__put",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chats"
        ],
        "summary": "Delete Chat",
        "operationId": "delete_chat_api_chats__chat_id__delete",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDeletedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/move": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Move Chat",
        "operationId": "move_chat_api_chats__chat_id__move_post",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/sandbox-policy": {
      "put": {
        "tags": [
          "chats"
        ],
        "summary": "Update Chat Sandbox Policy",
        "operationId": "update_chat_sandbox_policy_api_chats__chat_id__sandbox_policy_put",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSandboxPolicyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/sessions": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "Create Chat Session",
        "operationId": "create_chat_session_api_chats__chat_id__sessions_post",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "List Chat Sessions",
        "operationId": "list_chat_sessions_api_chats__chat_id__sessions_get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionInfo"
                  },
                  "title": "Response List Chat Sessions Api Chats  Chat Id  Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "operationId": "send_message_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendChatMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Chat": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "sandbox_sharing": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "isolated",
                  "session-shared",
                  "project-shared"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Sharing"
          },
          "sandbox_idle_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Idle Timeout Seconds"
          },
          "sandbox_resource_overrides": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Resource Overrides"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "display_name",
          "provider",
          "model",
          "created_at",
          "updated_at"
        ],
        "title": "Chat",
        "description": "A lightweight, non-agentic chat thread. `project_id` is optional; a standalone chat\nhas no meaningful sandbox policy until moved into a project."
      },
      "ChatApiMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatApiMessage",
        "description": "One turn's worth of chat history \u2014 matches `shared-types.ts`'s\n`ChatApiMessage` and `chat/messages.py::from_lc_messages`'s output\nexactly. Used both by `POST /api/chat`'s reply and\n`GET /api/sessions/{id}/messages`'s history."
      },
      "ChatDeletedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "chat_id"
        ],
        "title": "ChatDeletedResponse"
      },
      "ChatMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "chat_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Title"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatMessageRequest"
      },
...
      "ChatSandboxPolicyUpdate": {
        "properties": {
          "sharing": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "isolated",
                  "session-shared",
                  "project-shared"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharing"
          },
          "idle_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Timeout Seconds"
          },
          "resource_overrides": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Overrides"
          }
        },
        "type": "object",
        "title": "ChatSandboxPolicyUpdate",
        "description": "Unset (`None`) fields are left unchanged. Stored regardless of whether\n`project_id` is currently set, but only takes effect once it is \u2014 see\n`Chat`'s docstring."
      },
...
      "CreateChatRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "CreateChatRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoveChatRequest": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "MoveChatRequest"
      },
...
      "SendChatMessageResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "message": {
            "$ref": "#/components/schemas/ChatApiMessage"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "session_id",
          "message"
        ],
        "title": "SendChatMessageResponse"
      },
...
      "SessionInfo": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "owner_type": {
            "type": "string",
            "enum": [
              "agent",
              "chat"
            ],
            "title": "Owner Type"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "sandbox_sharing": {
            "type": "string",
            "enum": [
              "isolated",
              "session-shared",
              "project-shared"
            ],
            "title": "Sandbox Sharing",
            "default": "isolated"
          },
          "attached_to_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attached To Session Id"
          },
          "linked_session_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Session Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "owner_type",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "SessionInfo"
      },
...
      "UpdateChatRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "title": "UpdateChatRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}